tweak the wording introducing the properties of the meta vocab
Previously the section said:

> Metadata `meta` properties
> These properties MUST…

I found it confusing as cursory readers could misunderstand that it
applies to all `meta` properties. This PR makes it clear that it applies
to the properties defined in the following sections, and aligns the
wording with the other vocabs.

Also, the PR clarifies a bit the language (properties can't "reference
an expression", subexpressions do; properties are not children of `meta`
elements, etc).