Change mention of HTML target element in MO section "Rendering EPUB Content Document Elements"
The definition of MO src and epub:textref attribute values says only that they may refer to part of an EPUB Content Document, with no mention of what fragment-referencing schemes are supported.

In this section ("Rendering EPUB Content Document Elements") we get into what behavior SHOULD be in the case of HTML target elements or SVG fragment identifiers.

Now instead of saying "target elements", we say "reference to a specific element via its ID". I believe this is what the intention was with "target elements".